golf/GOLF_pipeline/requirements.txt
opencv-python
pyyaml
#pyrealsense2
mediapipe
pandas
numpy
